[www] Remove navigation bar
Move links to information card
