+++ packages/www/components/Common/ConsoleSection.module.css
@@ -1,15 +1,10 @@
-@media screen and (max-width: 600px) {
-  .Section {
-    margin: 0.5em;
-  }
-}
-
-@media screen and (min-width: 600px) {
-  .Section {
-    margin: 1em;
-  }
-}
-
 .Title {
+  position: sticky;
+  z-index: 1;
+  top: 0;
+  width: 100%;
+  padding: 1em 0;
+  text-align: center;
+  background-color: var(--global-background-color);
   font-weight: 500;
 }
