+++ packages/www/components/StickyCodeBlock.module.css
@@ -1,7 +1,4 @@
 .Block {
-  --code-block-width: 450px;
-  width: var(--code-block-width);
-  height: 670px;
   padding: 0.5em;
   line-height: 1.5em;
   margin: 0 auto;
@@ -10,8 +7,9 @@
 @media screen and (min-width: 960px) {
   .Block {
     position: fixed;
-    top: calc(50% - 700px / 2 + 32px);
-    left: calc(600px / 2 - var(--code-block-width) / 2);
+    top: 50%;
+    transform: translateY(-50%) translateX(-50%);
+    left: calc(600px / 2);
   }
 }
