+++ packages/www/pages/index.css
@@ -1,3 +1,7 @@
+:root {
+  --global-background-color: #f7f7f7;
+}
+
 body {
   margin: 0;
   padding: 0;
@@ -5,7 +9,7 @@ body {
     'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background: #f7f7f7;
+  background: var(--global-background-color);
   color: #38484f;
   overflow-x: hidden;
 }
